Update README.md
Building and Deploying the Soroban Token Smart Contract:*
This step involves building and deploying the smart contracts that represent the tokenized asset that users can send and receive. This is done by cloning the Soroban examples repository, navigating to the token directory, building the smart contract, and deploying it to Futurenet.

*Initializing a Token Contract:*
With the contracts deployed, the token contract needs to be initialized. This involves setting the administrator account, decimal precision, token name, and token symbol.

*Minting Tokens:*
After initializing the token contract, tokens need to be minted to the sender's account. This is done by invoking the mint function on the token contract.